refactor(agent): 迁移工作流核心到 workflow 根包并删除桥接层
- 将 workflow_agent.go 迁移到 workflow/engine.go
- 按职责拆分为多个文件: retry.go, condition_exec.go, join_edges.go, node_exec.go
- 更新包名为 workflow，添加子包 import
- 删除 types_bridge.go 桥接文件
- internal/agent/*.go 已清空，所有代码迁移完成
